Supports signed differences in the cache format. This uses the ZigZagEncoding & Decoding to convert signed integers to unsigned integers and vice versa essentially using the lsb to represent sign instead of msb. While encoding the signed differences are first converted to unsigned differences & then RLEd into the cache. While decoding the unsigned differences are RL decoded from the cache and converted to signed differences.
A new option to sort the features disregarding order of occurrence is provided to allow better compression.

Also fixes a bug in compressed option. Currently, reading from stdin doesn't work with compressed option turned on. So we check this and bail out with an error message. This is an effect of using zlib.