autotools_fixups_2
Ensure that the various require autotools files are distributed; all
files needed to build a distribution (aka "make dist") are present.

Update travis.yml to account for new build style.

C++11 is optional, C++0x is mandatory. C++11 is a feature superset, so
fall back to C++0x when necessary.