prefer_nullptr
Fix compile error in multilabel.cc: '\0' is not a pointer type. It is a
character and shouldn't be used as a subsitute for NULL, even if that
was previously valid in C.

Prefer the 'nullptr' keyword over 'NULL' because it is more type safe
than NULL (0). 'nullptr' was introduced as part of the C++0x standard;
VW requires at least C++0x support to compile.