call VW from R
This is a simple function to call VW from R, so far for logistic regression. Also it allows to use perf to compute the AUC on a validation set. I've found it a simple way to tune parameters based on a holdout set. I know vw has it's own methods for validation but I believe this is a common task. It's not really a new feature for VW but a simple function to facilitate validating models based on AUC. 
It's also fairly extendable to other problems.
I have written a reproducible example to show how it works. I'll add it as another file.