In bfgs.cc: (1) preconditioner is now recomputed every pass (2) Added reverting weight computation in test-only mode with input_feature_regularizer
In pylibvw.cc:
Added a function that returns the revert_weight field of an example

In loss_function.cc and loss_function.h:
Added a finalize_reverting_weight method for each loss function.
For log-loss and squared loss, this method computes the approximate reverting weight based on incremental BFGS.
For other loss functions, this method returns 0 and outputs a warning message that it is not implemented yet.