don't ignore the last line of RunTests
The last line of RunTests was ignored, even if it ended with a newline,
because eof(DATA) is true in that case.
If the last line contained e.g.
  train-sets/ref/active_cover_oracular.stderr
this stderr was not tested.
In this case, the file is missing, so the poor guy,
who tries to add a new test will got an error.