nn.cc: Add human readable named for hashes (#1179)
In Neural networks, some of the hidden networks had weights but there
was no hash associated to the weight. This made it difficult to read in
the Human readable format because the names were clashing (All the names
for the weights into the output layer had no space_name and hence the
name was always empty). This commit adds names for these weights which
are not seen.