add --classweight
$ vw --classweight LABEL1:WEIGHT1[,LABEL2:WEIGHT2...]

Multiplies the weight of the examples with label equal to LABEL1 by WEIGHT1, examples with label equal to LABEL2 by WEIGHT2 and so on.

This is already possible to achieve by rewriting the input file, however, making this built-in makes it much faster to experiment with different class weights.

Note, that this does affect the 'average loss' reported, the same way as modifying the examples' weight would.