Fix release build and unix symbol visibility. Add LTO mode.
Release builds were not working for two reasons.
First, it would always override -DCMAKE_BUILD_TYPE with "Debug"
Second, it would use CONFIG, which is not set on OSX with cmake 3.13.1 at least.

Finally, set the default symbol visibility to hidden on unix - linux is a misnomer in the build script...
Without it, we export something like 700 extra symbols that corresponds to all public functions in VW.
Exporting them on unix means making them available to be interposed and requiring an indirect call to be used
at all call sites.

The LTO mode, which can be used with -DLTO=1, enables clang's thinlto mode.