[OSX] Improve support for static linking.
On OSX, it's not possible to statically link against system libraries as they are closed source and Apple doesn't provide `.a` versions of them.

So we do the best thing we can given those constraints. We fix the build and eliminate all dependencies we can.
The resulting binaries should work on factory reset OSX machines, which is as good as we can aim for.