Fix race condition in daemon mode between parsing and inference threads. Fixes #2201 (#2283)
Once the parsing thread detects end-of-stream for the current socket, it triggers parser.cc::reset_source,
which closes the socket and accepts a new one.

Before it does so, it waits for the example queue to be empty. Though this is a necessary condition,
it's not enough as the last example will be in flight and the inference thread can run into a closed socket
when it tries to write to it.

The fix is to keep counts of both parsed and finished examples, then wait for them to be equal before closing
the current socket.