Simplify DFtoVW py api
This PR goal is to simplify the API of DFtoVW (issue #2458)

The simplifications made are the following:

    The user does not have to use the Col type anymore, although the class is still used internally (under the name _Col). Whether the argument passed refers to a dataframe column or is aconstant depends on an {attribute_name}_from_df attribute. The latter attribute has default value for the most common use case.
    A features parameter is added to the DFtoVW constructor. This parameter create a naked Namespace (namespace without any name or scaling factor). Of course, the user can still pass Namespace object(s) to the parameter namespaces for more advanced usage (but he can't use both namespaces and features in the same time).
    changed the main method name from DFtoVW.process_df to DFtoVW.convert_df which is more explicit in my opinion.

Now the DFtoVW can be used as follow:

df = pd.DataFrame({"y": [1], "a": [2], "b": [3]})
conv = DFtoVW(df=df, label=SimpleLabel("y"), features=[Feature(col) for col in ["a", "b"]])
conv.convert_df()

Implementation details
Using descriptor class

Along with the API simplifications I rewrite the internals so type and/or value checking can be performed directly on the attribute (or on the underlying column in the dataframe if the attribute refers to a column).

This type and value checking is performed using a descriptor class named AttributeDesciptor. The class will manage the setting of the attributes of the following classes: SimpleLabel, Feature, and Tag. When an attribute refers to a column (i.e when {attribute_name}_from_df=True) the expected type and value are passed to the _Col instance and checking is done when initializing DFtoVW (at this time the dataframe and thus the column is available).
Minor change

Some minor changes:

    the Namespace does not rely on AttributeDescriptor since none of its attributes is either a value or a column.
    descriptor and the managed class inherit from object to comply with python 2.7 (see Descriptor HowTo Guide)
    I create a _Tag type, but the parameter tag_from_df is available in DFtoVW and not in _Tag constructor since _Tag won't be used directly by the user.
    use black for style formatting with 79 line length.