[cbify] Fix dangling pointer in example due prediction array caching.
When memory reuse is done, we must ensure the invariant at any time, there's
only one valid "owning" pointer to that memory.

Owning, in this case, is a bit loose and doesn't mean smart pointer. It's
any pointer that could possibly be used to free the underlying memory.

In this case, cbify doesn't zero `example::pred` after it reclaims that v_array.