[Prediction] Modernize polypred a little by introducing ctor. (#2775)
* [Prediction] Modernize polypred a little by introducing ctor.

In order to introduce a ctor, and pave the way for a dtor, we need to make the
type copy/assignment safe.

Since v_array is not copy safe, so is polyprediction. So I decided to make it
not copyable/assignable, only moveable.

This change introduce a new guard, stash_guard, that borrows the name from the same name
git command. It's used all over the place to replace code that manually save/restore example::pred.

With this change in, it should be possible to move cleanup from ~example() to ~polyprediction.

* Fix fmt

* Don't use aggregate initializers :(

* Add unit tests for move semantics stuff in example.

Add a couple of v_array helpers v_move and v_extract that helps dealing with move semantics.

* try to fix linux.

* Fix cats failure.

Co-authored-by: Jack Gerrits <jackgerrits@users.noreply.github.com>