fix: Multiline Example handling when in multi-pass mode (#3143)
The logic for assembling and dispatching multiline examples was not handling end-of-pass examples correctly: When they are empty, they did not properly complete the multi_ex object and dispatch it to learn_multi_ex. Instead, it silently finished the end-of-pass example. Then when examples came in from a previous pass, they get appended to the inflight multi_ex, resulting in a malformed example in many cases (e.g. for cb_adf, this results in an example with multiple labeled actions)

* Fix multi_ex handling during multi-pass
* Create test for cb_adf multipass crash
* Add multipass cb tests