refactor: Migrate raw function pointers to std::function (#4461)
This PR replaces several raw function pointers with std::function objects.

The biggest changes are to the learner class, which is now no longer templated. To erase the data type, learner builders will bind the data object to their input functions via a lambda capture. To erase the example type, the lambda will accept a polymorphic_ex wrapper around VW::example* and VW::multi_ex*, which implicitly converts back to the original type. The resulting lambda can be stored in a std::function object with a type which does not depend on DataT nor ExampleT. This removes all undefined behavior around function pointer typecasting.

This PR also introduces the new term "bottom learner" to refer to the bottom-most learner of the reduction stack. This clears up the meaning of the word base, which now refers exclusively to the learner immediately below a given learner.