fix: early termination in fb tests due to misaligned read
The previous alignment fix for read did not resolve the special case where a filled buffer already has a misaligned fb data block somewhere in the middle. This would result in a properly-aligned fill(), followed a one or more successful reads, and then a failing one before the next fill().

The fix here is to ensure we try to shift-left when we encounter an unaligned normal (non-IO) buf_read call.

Also:
* added clarification in read_to that it does not support desired_align APIs (because it treats input as a byte stream, making alignment trivial)
* fixed semantics of error reporting out of fb_parser to correctly interact with parse_dispatch_loop, but also not report spurious failures