fix: desired_align test was broken in windows/debug
The reason for odd pass/fail behaviour across the CI is that MSVC introduces additional code in debug mode (or fails to apply certain optimizations, I did not fully investigate) which leads it to use a move on return of the test object (introduced when we removed the copy constructor).

Because Linux was optimizing that away in all cases, the test runs in CI (or locally) did not pick up the double-free bug which the previous commit (5deef6bb6bbcaff8e7d2c18eddf2631431b8e23e) fixed - an issue that was discovered on Windows[Debug].

The issue now is that the move constructor failed to copy the value of p.