fix: memory leak in run_parse_and_verify_test
In order to simulate enough of the parse dispatch loop to get the flatbuffer_parser to read multiple examples, we need to "dispatch" the examples after each pass through the parser, othrwise it is going to attempt to write in place over a single example.

When dispatching was introduced in the tests, it was used to swallow an extraneous newline example (which is used and swallowed internally by the example dispatcher to collect examples into multi_ex for, e.g. cb_adf), and without cleanup it leaks.