feat: read_span_flatbuffer() memory use and correctness fixes (#4684)
* read_span_flatbuffer: support cleaning up spare examples correctly

Consumers of VW as a library can provide their own event pools, etc. Previous parsers were always able to predict when an even would be needed ahead of time, so would only allocate when necessary. This was done by relying on a single incoming event preallocation to let the external host deallocate in the case of nothing to be parsed.

This does not work for the FB parser due to how it handles re-entrancy, and we do not want to spend the time re-architecting it to avoid this. The fix, in this case, is to expand the API to include a callback to return spare events back to the host's event pool.

* support header for using VW and RL api_status.h side-by-side
* Add additional error reporting to read_span_flatbuffer

* Reset parser when re-entering after bad parse

When re-using the flatbuffer parser across multiple invocations, the parser state could become invalid (retain references to deleted objects)

* Add more tests for bad inputs
* Add comments about what is going on in read_span_flatbuffer
* Fix a place where the parser was returning the semantically incorrect error code
* Remove dead code